+++ gutterball/src/main/java/org/candlepin/gutterball/eventhandler/ConsumerHandler.java
@@ -14,57 +14,20 @@
  */
 package org.candlepin.gutterball.eventhandler;
 
-import org.apache.commons.lang.StringUtils;
-import org.bson.BSONCallback;
-
-import org.candlepin.gutterball.bsoncallback.DateAndEscapeCallback;
 import org.candlepin.gutterball.curator.ConsumerCurator;
-import org.candlepin.gutterball.model.Event;
+import org.candlepin.gutterball.model.ProcessedEvent;
 
 import com.google.inject.Inject;
-import com.mongodb.DBObject;
-import com.mongodb.util.JSON;
 
 /**
  * ConsumerHandler to properly update the database when
  * a consumer based event is received
  */
 @HandlerTarget("CONSUMER")
-public class ConsumerHandler implements EventHandler {
-
-    protected ConsumerCurator consumerCurator;
-    protected BSONCallback callback;
+public class ConsumerHandler extends AbstractBaseHandler<ProcessedEvent, ConsumerCurator> {
 
     @Inject
     public ConsumerHandler(ConsumerCurator consumerCurator) {
-        this.consumerCurator = consumerCurator;
-        callback = new DateAndEscapeCallback();
-    }
-
-    public void handleEvent(Event event) {
-        String newEntityJson = (String) event.get("newEntity");
-        if (!StringUtils.isBlank(newEntityJson)) {
-            DBObject newEntity = (DBObject) JSON.parse(newEntityJson, callback);
-            // TODO: we should have a consumer record that we maintain with
-            // created/deleted/owner info to narrow our queries.
-            // Each consumer record could also store that records id in some
-            // sort of _master_id field.
-            consumerCurator.insert(newEntity);
-        }
-    }
-
-    @Override
-    public void handleCreated(Event event) {
-        handleEvent(event);
-    }
-
-    @Override
-    public void handleUpdated(Event event) {
-        handleEvent(event);
-    }
-
-    @Override
-    public void handleDeleted(Event event) {
-        // Eventually we'll need to update some sort of consumer master record
+        super(consumerCurator);
     }
 }
