+++ gutterball/src/main/java/org/candlepin/gutterball/eventhandler/EntitlementHandler.java
@@ -14,17 +14,10 @@
  */
 package org.candlepin.gutterball.eventhandler;
 
-import org.candlepin.gutterball.bsoncallback.DateAndEscapeCallback;
 import org.candlepin.gutterball.curator.EntitlementCurator;
-import org.candlepin.gutterball.model.Event;
+import org.candlepin.gutterball.model.ProcessedEvent;
 
 import com.google.inject.Inject;
-import com.mongodb.DBObject;
-import com.mongodb.util.JSON;
-
-import org.apache.commons.lang.StringUtils;
-import org.bson.BSONCallback;
-
 /**
  * GuestIdHandler class to deal with guestId events
  *
@@ -36,41 +29,10 @@ import org.bson.BSONCallback;
  * behind our backs.
  */
 @HandlerTarget("ENTITLEMENT")
-public class EntitlementHandler implements EventHandler {
-
-    protected EntitlementCurator curator;
-    protected BSONCallback callback;
+public class EntitlementHandler extends AbstractBaseHandler<ProcessedEvent, EntitlementCurator> {
 
     @Inject
     public EntitlementHandler(EntitlementCurator curator) {
-        this.curator = curator;
-        callback = new DateAndEscapeCallback();
-    }
-
-    public void handleEvent(Event event) {
-        String newEntityJson = (String) event.get("newEntity");
-        if (!StringUtils.isBlank(newEntityJson)) {
-            DBObject newEntity = (DBObject) JSON.parse(newEntityJson, callback);
-            // TODO: we should have a record that we maintain with
-            // created/deleted/owner info to narrow our queries.
-            // Each consumer record could also store that records id in some
-            // sort of _master_id field.
-            curator.insert(newEntity);
-        }
-    }
-
-    @Override
-    public void handleCreated(Event event) {
-        handleEvent(event);
-    }
-
-    @Override
-    public void handleUpdated(Event event) {
-        handleEvent(event);
-    }
-
-    @Override
-    public void handleDeleted(Event event) {
-        // Eventually we'll need to update some sort of consumer master record
+        super(curator);
     }
 }
